django_ecommerce/requirements.txt
decorator==4.0.10
Django==1.8.2
django-debug-toolbar==1.6
djangorestframework==3.5.2
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
mock==1.0.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
psycopg2==2.6.2
ptyprocess==0.5.1
Pygments==2.1.3
requests==2.3.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
stripe==1.9.2
traitlets==4.3.1
wcwidth==0.1.7
